Split export dialog sections
